pyproject.toml
[project]
name = "ui-module"
version = "0.1.0"
description = "MCP-first Server-Driven UI Framework"
authors = [
    {name = "Antigravity", email = "antigravity@example.com"},
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[project.scripts]
ui-module = "ui_module.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ui_module"]
